Test
Try to compile under VS2010.

+++ Client/MAttackZoneEffectGenerator.cpp
@@ -110,7 +110,7 @@ MAttackZoneEffectGenerator::Generate( const EFFECTGENERATOR_INFO& egInfo )
 			cy = sy - ty;
 		}
 
-		int currentPixel = sqrt(cx*cx + cy*cy);
+		int currentPixel = sqrt((double)cx*cx + cy*cy);
 
 		//float basis = ((cx==0)? 0 : (float)cy / (float)cx);
